api_server/requirements.txt
Django>=5.1
djangorestframework>=3.15
psycopg2-binary>=2.9
django-cors-headers>=4.4
Pillow>=11.0
python-dotenv>=1.0
django-jazzmin>=3.0.1
gunicorn>=23.0
whitenoise>=6.8
dj-database-url>=2.3
django-cloudinary-storage>=0.3.0
cloudinary>=1.41.0
